Split core testsuite so that all Cache CR tests belong in test/e2e/cache
