+++ Makefile
@@ -55,6 +55,11 @@ unit-test: manager
 test: manager manifests
 	scripts/run-tests.sh main
 
+.PHONY: cache-test
+## Execute end to end (e2e) tests for Cache CRs
+cache-test: manager manifests
+	scripts/run-tests.sh cache
+
 .PHONY: multinamespace-test
 ## Execute end to end (e2e) tests in multinamespace mode
 multinamespace-test: manager manifests
