+++ test/e2e/main/main_test.go
@@ -18,7 +18,6 @@ import (
 	"github.com/iancoleman/strcase"
 	ispnv1 "github.com/infinispan/infinispan-operator/api/v1"
 	v1 "github.com/infinispan/infinispan-operator/api/v1"
-	"github.com/infinispan/infinispan-operator/api/v2alpha1"
 	"github.com/infinispan/infinispan-operator/controllers"
 	cconsts "github.com/infinispan/infinispan-operator/controllers/constants"
 	hash "github.com/infinispan/infinispan-operator/pkg/hash"
@@ -385,7 +384,7 @@ func checkRestConnection(hostAddr string, client tutils.HTTPClient) {
 	tutils.ExpectNoError(err)
 	defer tutils.CloseHttpResponse(rsp)
 	if rsp.StatusCode != http.StatusOK {
-		panic(httpError{rsp.StatusCode})
+		tutils.ThrowHTTPError(rsp)
 	}
 }
 
@@ -511,7 +510,7 @@ func testClientCert(t *testing.T, initializer func(*v1.Infinispan) (v1.ClientCer
 
 	// Ensure that we can connect to the endpoint with TLS
 	host, client := tutils.HTTPSClientAndHost(spec, tlsConfig, testKube)
-	createCacheAndValidate("test", host, "", client)
+	tutils.CreateCacheAndValidate("test", host, "", client)
 }
 
 // Test if spec.container.cpu update is handled
@@ -681,13 +680,13 @@ func testCacheService(testName string) {
 	hostAddr, client := tutils.HTTPClientAndHost(ispn, testKube)
 
 	cacheName := "default"
-	waitForCacheToBeCreated(cacheName, hostAddr, client)
+	tutils.WaitForCacheToBeCreated(cacheName, hostAddr, client)
 
 	key := "test"
 	value := "test-operator"
-	keyURL := fmt.Sprintf("%v/%v", cacheURL(cacheName, hostAddr), key)
-	putViaRoute(keyURL, value, client)
-	actual := getViaRoute(keyURL, client)
+	keyURL := fmt.Sprintf("%v/%v", tutils.CacheURL(cacheName, hostAddr), key)
+	tutils.PutViaRoute(keyURL, value, client)
+	actual := tutils.GetViaRoute(keyURL, client)
 
 	if actual != value {
 		panic(fmt.Errorf("unexpected actual returned: %v (value %v)", actual, value))
@@ -703,20 +702,20 @@ func TestPermanentCache(t *testing.T) {
 	// Define function for the generic stop/start test procedure
 	var createPermanentCache = func(ispn *ispnv1.Infinispan) {
 		hostAddr, client := tutils.HTTPClientAndHost(ispn, testKube)
-		createCacheAndValidate(cacheName, hostAddr, "PERMANENT", client)
+		tutils.CreateCacheAndValidate(cacheName, hostAddr, "PERMANENT", client)
 	}
 
 	var usePermanentCache = func(ispn *ispnv1.Infinispan) {
 		hostAddr, client := tutils.HTTPClientAndHost(ispn, testKube)
 		key := "test"
 		value := "test-operator"
-		keyURL := fmt.Sprintf("%v/%v", cacheURL(cacheName, hostAddr), key)
-		putViaRoute(keyURL, value, client)
-		actual := getViaRoute(keyURL, client)
+		keyURL := fmt.Sprintf("%v/%v", tutils.CacheURL(cacheName, hostAddr), key)
+		tutils.PutViaRoute(keyURL, value, client)
+		actual := tutils.GetViaRoute(keyURL, client)
 		if actual != value {
 			panic(fmt.Errorf("unexpected actual returned: %v (value %v)", actual, value))
 		}
-		deleteCache(cacheName, hostAddr, client)
+		tutils.DeleteCache(cacheName, hostAddr, client)
 	}
 
 	genericTestForGracefulShutdown(name, createPermanentCache, usePermanentCache)
@@ -736,19 +735,19 @@ func TestCheckDataSurviveToShutdown(t *testing.T) {
 	// Define function for the generic stop/start test procedure
 	var createCacheWithFileStore = func(ispn *ispnv1.Infinispan) {
 		hostAddr, client := tutils.HTTPClientAndHost(ispn, testKube)
-		createCacheWithXMLTemplate(cacheName, hostAddr, template, client)
-		keyURL := fmt.Sprintf("%v/%v", cacheURL(cacheName, hostAddr), key)
-		putViaRoute(keyURL, value, client)
+		tutils.CreateCacheWithXMLTemplate(cacheName, hostAddr, template, client)
+		keyURL := fmt.Sprintf("%v/%v", tutils.CacheURL(cacheName, hostAddr), key)
+		tutils.PutViaRoute(keyURL, value, client)
 	}
 
 	var useCacheWithFileStore = func(ispn *ispnv1.Infinispan) {
 		hostAddr, client := tutils.HTTPClientAndHost(ispn, testKube)
-		keyURL := fmt.Sprintf("%v/%v", cacheURL(cacheName, hostAddr), key)
-		actual := getViaRoute(keyURL, client)
+		keyURL := fmt.Sprintf("%v/%v", tutils.CacheURL(cacheName, hostAddr), key)
+		actual := tutils.GetViaRoute(keyURL, client)
 		if actual != value {
 			panic(fmt.Errorf("unexpected actual returned: %v (value %v)", actual, value))
 		}
-		deleteCache(cacheName, hostAddr, client)
+		tutils.DeleteCache(cacheName, hostAddr, client)
 	}
 
 	genericTestForGracefulShutdown(name, createCacheWithFileStore, useCacheWithFileStore)
@@ -807,14 +806,14 @@ func TestExternalService(t *testing.T) {
 	hostAddr, client := tutils.HTTPClientAndHost(ispn, testKube)
 
 	cacheName := "test"
-	createCacheAndValidate(cacheName, hostAddr, "", client)
-	defer deleteCache(cacheName, hostAddr, client)
+	tutils.CreateCacheAndValidate(cacheName, hostAddr, "", client)
+	defer tutils.DeleteCache(cacheName, hostAddr, client)
 
 	key := "test"
 	value := "test-operator"
-	keyURL := fmt.Sprintf("%v/%v", cacheURL(cacheName, hostAddr), key)
-	putViaRoute(keyURL, value, client)
-	actual := getViaRoute(keyURL, client)
+	keyURL := fmt.Sprintf("%v/%v", tutils.CacheURL(cacheName, hostAddr), key)
+	tutils.PutViaRoute(keyURL, value, client)
+	actual := tutils.GetViaRoute(keyURL, client)
 
 	if actual != value {
 		panic(fmt.Errorf("unexpected actual returned: %v (value %v)", actual, value))
@@ -929,14 +928,14 @@ func testAuthentication(ispn *ispnv1.Infinispan, schema, usr, pass string) {
 
 	cacheName := "test"
 	createCacheBadCreds(cacheName, hostAddr, badClient)
-	createCacheAndValidate(cacheName, hostAddr, "", client)
-	defer deleteCache(cacheName, hostAddr, client)
+	tutils.CreateCacheAndValidate(cacheName, hostAddr, "", client)
+	defer tutils.DeleteCache(cacheName, hostAddr, client)
 
 	key := "test"
 	value := "test-operator"
-	keyURL := fmt.Sprintf("%v/%v", cacheURL(cacheName, hostAddr), key)
-	putViaRoute(keyURL, value, client)
-	actual := getViaRoute(keyURL, client)
+	keyURL := fmt.Sprintf("%v/%v", tutils.CacheURL(cacheName, hostAddr), key)
+	tutils.PutViaRoute(keyURL, value, client)
+	actual := tutils.GetViaRoute(keyURL, client)
 
 	if actual != value {
 		panic(fmt.Errorf("unexpected actual returned: %v (value %v)", actual, value))
@@ -974,7 +973,7 @@ func TestAuthenticationDisabled(t *testing.T) {
 	rsp, err := client.Get(url, nil)
 	tutils.ExpectNoError(err)
 	if rsp.StatusCode != http.StatusOK {
-		panic(httpError{rsp.StatusCode})
+		tutils.ThrowHTTPError(rsp)
 	}
 }
 
@@ -1000,7 +999,7 @@ func TestAuthorizationDisabledByDefault(t *testing.T) {
 		rsp, err := client.Get(url, nil)
 		tutils.ExpectNoError(err)
 		if rsp.StatusCode != http.StatusOK {
-			panic(httpError{rsp.StatusCode})
+			tutils.ThrowHTTPError(rsp)
 		}
 	}
 	testAuthorization(ispn, identities, verify)
@@ -1043,11 +1042,11 @@ func TestAuthorizationWithCustomRoles(t *testing.T) {
 	}
 
 	verify := func(hostAddr string, client tutils.HTTPClient) {
-		createCacheAndValidate("succeed-cache", hostAddr, "", client)
+		tutils.CreateCacheAndValidate("succeed-cache", hostAddr, "", client)
 		schema := testKube.GetSchemaForRest(ispn)
-		rsp := createCache("fail-cache", hostAddr, "", tutils.NewHTTPClient("monitor-user", "pass", schema))
+		rsp := tutils.CreateCache("fail-cache", hostAddr, "", tutils.NewHTTPClient("monitor-user", "pass", schema))
 		if rsp.StatusCode != http.StatusForbidden {
-			panic(httpError{rsp.StatusCode})
+			tutils.ThrowHTTPError(rsp)
 		}
 	}
 	testAuthorization(ispn, identities, verify)
@@ -1223,173 +1222,18 @@ func prepareWebServer() *corev1.ConfigMap {
 	return webServerConfig
 }
 
-func cacheURL(cacheName, hostAddr string) string {
-	return fmt.Sprintf("%v/rest/v2/caches/%s", hostAddr, cacheName)
-}
-
-type httpError struct {
-	status int
-}
-
-func (e *httpError) Error() string {
-	return fmt.Sprintf("unexpected response %v", e.status)
-}
-
-func createCache(cacheName, hostAddr string, flags string, client tutils.HTTPClient) *http.Response {
-	httpURL := cacheURL(cacheName, hostAddr)
-	headers := map[string]string{}
-	if flags != "" {
-		headers["Flags"] = flags
-	}
-	resp, err := client.Post(httpURL, "", headers)
-	tutils.ExpectNoError(err)
-	return resp
-}
-
-func createCacheAndValidate(cacheName, hostAddr string, flags string, client tutils.HTTPClient) {
-	resp := createCache(cacheName, hostAddr, flags, client)
-	defer tutils.CloseHttpResponse(resp)
-	if resp.StatusCode != http.StatusOK {
-		panic(httpError{resp.StatusCode})
-	}
-}
-
 func createCacheBadCreds(cacheName, hostAddr string, client tutils.HTTPClient) {
 	defer func() {
 		data := recover()
 		if data == nil {
 			panic("createCacheBadCred should fail, but it doesn't")
 		}
-		err := data.(httpError)
-		if err.status != http.StatusUnauthorized {
+		err := data.(tutils.HttpError)
+		if err.Status != http.StatusUnauthorized {
 			panic(err)
 		}
 	}()
-	createCacheAndValidate(cacheName, hostAddr, "", client)
-}
-
-func createCacheWithXMLTemplate(cacheName, hostAddr, template string, client tutils.HTTPClient) {
-	httpURL := cacheURL(cacheName, hostAddr)
-	fmt.Printf("Create cache: %v\n", httpURL)
-	headers := map[string]string{
-		"Content-Type": "application/xml;charset=UTF-8",
-	}
-	resp, err := client.Post(httpURL, template, headers)
-	defer tutils.CloseHttpResponse(resp)
-	tutils.ExpectNoError(err)
-	// Accept all the 2xx success codes
-	if resp.StatusCode < http.StatusOK || resp.StatusCode >= http.StatusMultipleChoices {
-		throwHTTPError(resp)
-	}
-}
-
-func deleteCache(cacheName, hostAddr string, client tutils.HTTPClient) {
-	httpURL := cacheURL(cacheName, hostAddr)
-	resp, err := client.Delete(httpURL, nil)
-	tutils.ExpectNoError(err)
-
-	if resp.StatusCode != http.StatusOK {
-		panic(httpError{resp.StatusCode})
-	}
-}
-
-func getViaRoute(url string, client tutils.HTTPClient) string {
-	resp, err := client.Get(url, nil)
-	tutils.ExpectNoError(err)
-	defer func(Body io.ReadCloser) {
-		tutils.ExpectNoError(Body.Close())
-	}(resp.Body)
-	if resp.StatusCode != http.StatusOK {
-		throwHTTPError(resp)
-	}
-	bodyBytes, err := ioutil.ReadAll(resp.Body)
-	tutils.ExpectNoError(err)
-	return string(bodyBytes)
-}
-
-func putViaRoute(url, value string, client tutils.HTTPClient) {
-	headers := map[string]string{
-		"Content-Type": "text/plain",
-	}
-	resp, err := client.Post(url, value, headers)
-	defer tutils.CloseHttpResponse(resp)
-	tutils.ExpectNoError(err)
-	if resp.StatusCode != http.StatusNoContent {
-		throwHTTPError(resp)
-	}
-}
-
-func waitForCacheToBeCreated(cacheName, hostAddr string, client tutils.HTTPClient) {
-	err := wait.Poll(tutils.DefaultPollPeriod, tutils.MaxWaitTimeout, func() (done bool, err error) {
-		httpURL := cacheURL(cacheName, hostAddr)
-		fmt.Printf("Waiting for cache to be created")
-		resp, err := client.Get(httpURL, nil)
-		if err != nil {
-			return false, err
-		}
-		return resp.StatusCode == http.StatusOK, nil
-	})
-	tutils.ExpectNoError(err)
-}
-
-func throwHTTPError(resp *http.Response) {
-	errorBytes, _ := ioutil.ReadAll(resp.Body)
-	panic(fmt.Errorf("unexpected HTTP status code (%d): %s", resp.StatusCode, string(errorBytes)))
-}
-
-func TestCacheCR(t *testing.T) {
-	t.Parallel()
-	spec := tutils.DefaultSpec(testKube)
-	name := strcase.ToKebab(t.Name())
-	spec.Name = name
-	spec.Labels = map[string]string{"test-name": t.Name()}
-	testKube.CreateInfinispan(spec, tutils.Namespace)
-	defer testKube.CleanNamespaceAndLogOnPanic(tutils.Namespace, spec.Labels)
-	testKube.WaitForInfinispanPods(1, tutils.SinglePodTimeout, spec.Name, tutils.Namespace)
-	ispn := testKube.WaitForInfinispanCondition(spec.Name, spec.Namespace, ispnv1.ConditionWellFormed)
-
-	//Test for CacheCR with Templatename
-
-	cacheCRTemplateName := createCacheWithCR("cache-with-static-template", spec.Namespace, name)
-	cacheCRTemplateName.Spec.TemplateName = "org.infinispan.DIST_SYNC"
-	testCacheWithCR(ispn, cacheCRTemplateName)
-
-	//Test for CacheCR with TemplateXML
-
-	cacheCRTemplateXML := createCacheWithCR("cache-with-xml-template", spec.Namespace, name)
-	cacheCRTemplateXML.Spec.Template = "<infinispan><cache-container><distributed-cache name=\"cache-with-xml-template\" mode=\"SYNC\"><persistence><file-store/></persistence></distributed-cache></cache-container></infinispan>"
-	testCacheWithCR(ispn, cacheCRTemplateXML)
-}
-
-func testCacheWithCR(ispn *ispnv1.Infinispan, cache *v2alpha1.Cache) {
-	key := "testkey"
-	value := "test-operator"
-	testKube.Create(cache)
-	hostAddr, client := tutils.HTTPClientAndHost(ispn, testKube)
-	condition := v2alpha1.CacheCondition{
-		Type:   "Ready",
-		Status: "True",
-	}
-	testKube.WaitForCacheCondition(cache.Spec.Name, cache.Namespace, condition)
-	waitForCacheToBeCreated(cache.Spec.Name, hostAddr, client)
-	testBasicCacheUsage(key, value, cache.Spec.Name, hostAddr, client)
-	defer testKube.DeleteCache(cache)
-}
-func createCacheWithCR(cacheName string, nameSpace string, clusterName string) *v2alpha1.Cache {
-	return &v2alpha1.Cache{
-		TypeMeta: metav1.TypeMeta{
-			APIVersion: "infinispan.org/v2alpha1",
-			Kind:       "Cache",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      cacheName,
-			Namespace: nameSpace,
-		},
-		Spec: v2alpha1.CacheSpec{
-			ClusterName: clusterName,
-			Name:        cacheName,
-		},
-	}
+	tutils.CreateCacheAndValidate(cacheName, hostAddr, "", client)
 }
 
 func TestPodDegradationAfterOOM(t *testing.T) {
@@ -1414,11 +1258,11 @@ func TestPodDegradationAfterOOM(t *testing.T) {
 		`"><encoding media-type="text/plain"/></replicated-cache></cache-container></infinispan>`
 	veryLongValue := GenerateStringWithCharset(6000)
 	hostAddr, client := tutils.HTTPClientAndHost(ispn, testKube)
-	createCacheWithXMLTemplate(cacheName, hostAddr, template, client)
+	tutils.CreateCacheWithXMLTemplate(cacheName, hostAddr, template, client)
 
 	//Generate tons of random entries
 	for key := 1; key < 50000; key++ {
-		keyURL := fmt.Sprintf("%v/%v", cacheURL(cacheName, hostAddr), key)
+		keyURL := fmt.Sprintf("%v/%v", tutils.CacheURL(cacheName, hostAddr), key)
 		headers := map[string]string{"Content-Type": "text/plain"}
 		_, error := client.Post(keyURL, veryLongValue, headers)
 
@@ -1508,17 +1352,7 @@ func testCustomConfig(t *testing.T, configMap *corev1.ConfigMap) {
 	key := "testkey"
 	value := "test-operator"
 	hostAddr, client := tutils.HTTPClientAndHost(ispn, testKube)
-	testBasicCacheUsage(key, value, t.Name(), hostAddr, client)
-}
-
-func testBasicCacheUsage(key, value, cacheName, hostAddr string, client tutils.HTTPClient) {
-	keyURL := fmt.Sprintf("%v/%v", cacheURL(cacheName, hostAddr), key)
-	putViaRoute(keyURL, value, client)
-	actual := getViaRoute(keyURL, client)
-
-	if actual != value {
-		panic(fmt.Errorf("unexpected actual returned: %v (value %v)", actual, value))
-	}
+	tutils.TestBasicCacheUsage(key, value, t.Name(), hostAddr, client)
 }
 
 // TestUserCustomConfigWithAuthUpdate tests that user custom config works well with update
@@ -1533,7 +1367,7 @@ func TestUserCustomConfigWithAuthUpdate(t *testing.T) {
 		key := "testkey"
 		value := "test-operator"
 		hostAddr, client := tutils.HTTPClientAndHost(ispn, testKube)
-		testBasicCacheUsage(key, value, t.Name(), hostAddr, client)
+		tutils.TestBasicCacheUsage(key, value, t.Name(), hostAddr, client)
 		ispn.Spec.Security.EndpointAuthentication = pointer.BoolPtr(true)
 	}
 	var verifier = func(ispn *ispnv1.Infinispan, ss *appsv1.StatefulSet) {
@@ -1542,7 +1376,7 @@ func TestUserCustomConfigWithAuthUpdate(t *testing.T) {
 		key := "testkey"
 		value := "test-operator"
 		hostAddr, client := tutils.HTTPClientAndHost(ispn, testKube)
-		testBasicCacheUsage(key, value, t.Name(), hostAddr, client)
+		tutils.TestBasicCacheUsage(key, value, t.Name(), hostAddr, client)
 	}
 	ispn := tutils.DefaultSpec(testKube)
 	ispn.Name = strcase.ToKebab(t.Name())
@@ -1564,7 +1398,7 @@ func TestUserCustomConfigUpdateOnNameChange(t *testing.T) {
 		key := "testkey"
 		value := "test-operator"
 		hostAddr, client := tutils.HTTPClientAndHost(ispn, testKube)
-		testBasicCacheUsage(key, value, t.Name(), hostAddr, client)
+		tutils.TestBasicCacheUsage(key, value, t.Name(), hostAddr, client)
 		configMapChanged := newCustomConfigMap(t.Name()+"Changed", "xml")
 		testKube.Create(configMapChanged)
 		ispn.Spec.ConfigMapName = configMapChanged.Name
@@ -1575,7 +1409,7 @@ func TestUserCustomConfigUpdateOnNameChange(t *testing.T) {
 		key := "testkey"
 		value := "test-operator"
 		hostAddr, client := tutils.HTTPClientAndHost(ispn, testKube)
-		testBasicCacheUsage(key, value, t.Name()+"Changed", hostAddr, client)
+		tutils.TestBasicCacheUsage(key, value, t.Name()+"Changed", hostAddr, client)
 	}
 	ispn := tutils.DefaultSpec(testKube)
 	ispn.Name = strcase.ToKebab(t.Name())
@@ -1597,7 +1431,7 @@ func TestUserCustomConfigUpdateOnChange(t *testing.T) {
 		key := "testkey"
 		value := "test-operator"
 		hostAddr, client := tutils.HTTPClientAndHost(ispn, testKube)
-		testBasicCacheUsage(key, value, t.Name(), hostAddr, client)
+		tutils.TestBasicCacheUsage(key, value, t.Name(), hostAddr, client)
 		configMapUpdated := newCustomConfigMap(newCacheName, "xml")
 		// Reuse old name to test CM in-place update
 		configMapUpdated.Name = strcase.ToKebab(t.Name())
@@ -1609,7 +1443,7 @@ func TestUserCustomConfigUpdateOnChange(t *testing.T) {
 		key := "testkey"
 		value := "test-operator"
 		hostAddr, client := tutils.HTTPClientAndHost(ispn, testKube)
-		testBasicCacheUsage(key, value, newCacheName, hostAddr, client)
+		tutils.TestBasicCacheUsage(key, value, newCacheName, hostAddr, client)
 	}
 	ispn := tutils.DefaultSpec(testKube)
 	ispn.Name = strcase.ToKebab(t.Name())
@@ -1636,7 +1470,7 @@ func TestUserCustomConfigUpdateOnAdd(t *testing.T) {
 		key := "testkey"
 		value := "test-operator"
 		hostAddr, client := tutils.HTTPClientAndHost(ispn, testKube)
-		testBasicCacheUsage(key, value, t.Name(), hostAddr, client)
+		tutils.TestBasicCacheUsage(key, value, t.Name(), hostAddr, client)
 	}
 	ispn := tutils.DefaultSpec(testKube)
 	ispn.Name = strcase.ToKebab(t.Name())
